requirements.txt
flask
twilio
PyMuPDF
sentence-transformers
faiss-cpu
openai
